macos: a list of its own, and a pane for what the server knows about a row
Two floating panels, deliberately parallel to the hover card rather than
subclassed from it. The hover card sets `becomesKeyOnlyIfNeeded` because
its text is selectable, and that one line is fatal here: clicking a row
would take key status, the caret would stop blinking and the text view
would stop receiving arrows — the list becoming un-navigable in the same
gesture that opened it. What the two share is the flip-and-clamp geometry,
and only that.

The list is a table rather than a stack of views: a language server answers
with hundreds of items. Its selection is drawn from the theme rather than
by the table, because a table's own selection renders in the inactive grey
in a window that never becomes key — which is this panel's permanent state.
Rows accept the first mouse, or the first click on one is swallowed as an
activation click and every acceptance takes two.

The documentation pane distinguishes four answers, and the distinction is
the whole feature: documentation present, documentation absent, the list it
was about having been superseded, and the server not offering resolution at
all. The middle two look identical if you only check for nil, and getting
them wrong means blanking the pane while the reader arrow-keys down a list
— a superseded answer is the *common* case, since each request cancels the
one before it. The fourth is Kotlin's permanent answer, so no spinner and
no retry: `kotlin-language-server` neither resolves nor ships per-item
documentation, and the pane is legitimately empty there forever.

Markdown is rendered as markdown only when the server said so. Running the
splitter over plain text eats horizontal rules and reflows paragraphs that
meant to keep their breaks.

Also here, three test-only defects that stopped the suite from running at
all, which is why they are fixed now rather than batched: an `#expect` over
a `rethrows` call that reads as throwing inside the macro's expansion, a
`Comment` built from a loop variable rather than a literal, and — the one
worth remembering — a float assertion comparing `CGFloat` against `Double`.
Ordinary Swift compares those through an implicit conversion; **that
conversion does not survive `#expect`'s expansion**, so the macro failed on
two operands that were bit-for-bit identical and reported it as an
arithmetic mismatch between 20.0 and 20.0.

+++ macos/Tests/Terminal/CodeCompletionPanelTests.swift
@@ -53,10 +53,26 @@ struct CodeCompletionPanelTests {
     /// The list's rhythm follows the file's, so a bigger editor font gives
     /// taller rows rather than a denser list.
     @Test func rowHeightFollowsTheEditorsFont() {
-        #expect(CodeCompletionPanel.rowHeight(for: font) == ceil(12 * 1.6))
+        /// Both sides are spelled `CGFloat` on purpose, and this is not
+        /// pedantry — it is the difference between this test passing and
+        /// failing.
+        ///
+        /// `rowHeight(for:)` returns `CGFloat`; a bare `ceil(12 * 1.6)`
+        /// infers `Double`. Ordinary Swift compares those happily through
+        /// the implicit `CGFloat`/`Double` conversion, but **that conversion
+        /// does not survive `#expect`'s expansion**, so the macro compares
+        /// two operands of different types and reports a failure between
+        /// values that are bit-for-bit identical:
+        ///
+        ///     measured 4626322717216342016 (CGFloat)
+        ///     expected 4626322717216342016 (Double)
+        ///
+        /// Which reads as an arithmetic bug and is not one. Any float
+        /// assertion in this suite has to name its type on both sides.
+        #expect(CodeCompletionPanel.rowHeight(for: font) == CGFloat(ceil(12 * 1.6)))
         #expect(
             CodeCompletionPanel.rowHeight(for: .monospacedSystemFont(ofSize: 18, weight: .regular))
-                == ceil(18 * 1.6)
+                == CGFloat(ceil(18 * 1.6))
         )
     }
